more comments + constants

+++ tonnfcclientandroid/src/main/java/com/tonnfccard/smartcard/CommonConstants.java
@@ -1,10 +1,10 @@
 package com.tonnfccard.smartcard;
 
 class CommonConstants {
-  final static byte SELECT_CLA = (byte) 0x00;
+  final static byte SELECT_CLA = 0x00;
   final static byte SELECT_INS = (byte) 0xA4;
-  final static byte SELECT_P1 = (byte) 0x04;
-  final static byte SELECT_P2 = (byte) 0x00;
+  final static byte SELECT_P1 = 0x04;
+  final static byte SELECT_P2 = 0x00;
   final static byte LE = 0x00;
   final static byte NEGATIVE_LE = (byte) -1;
   final static String SELECT_COIN_MANAGER_APDU_NAME = "SELECT_COIN_MANAGER";
